Move `Colour` class to `yaml_checker.py`
Fixes #304
