+++ httomo/yaml_checker.py
@@ -9,7 +9,6 @@ import yaml
 
 from pathlib import Path
 
-from httomo.utils import Colour
 from httomo.yaml_utils import get_packages_current_version
 from httomo.ui_layer import (
     get_regex_pattern,
@@ -25,6 +24,25 @@ __all__ = [
 
 MethodConfig: TypeAlias = Dict[str, Any]
 PipelineConfig: TypeAlias = List[MethodConfig]
+
+
+class Colour:
+    """
+    Class for storing the ANSI escape codes for different colours.
+    """
+
+    LIGHT_BLUE = "\033[1;34m"
+    LIGHT_BLUE_BCKGR = "\033[1;44m"
+    BLUE = "\33[94m"
+    CYAN = "\33[96m"
+    GREEN = "\33[92m"
+    YELLOW = "\33[93m"
+    MAGENTA = "\33[95m"
+    RED = "\33[91m"
+    END = "\033[0m"
+    BVIOLET = "\033[1;35m"
+    LYELLOW = "\033[33m"
+    BACKG_RED = "\x1b[6;37;41m"
 
 
 def sanity_check(conf_generator: Iterator[Any]) -> bool:
